requirements.txt
requests==2.31.0
spacy==3.7.1
boto3==1.28.62
asgiref==3.7.2
botocore==1.31.62
python-dotenv==1.0.0
telebot~=0.0.5